Refactored Login and Common components.
