+++ src/components/common/NavBar/NavBar.styles.js
@@ -1,0 +1,10 @@
+const styles = {
+    root: {
+      display: "flex"
+    },
+    menuIcon: {
+      marginLeft: "auto"
+    }
+  };
+
+export default styles;
