+++ src/components/common/NavMenu/NavMenu.test.js
@@ -1,10 +1,10 @@
 import React from "react";
 import { shallow } from "enzyme";
-import NavBar from "../NavBar";
+import NavMenu from "./NavMenu";
 
 describe("<NavMenu />", () => {
   test("it renders correctly", () => {
-    let wrapper = shallow(<NavBar />).dive();
+    let wrapper = shallow(<NavMenu />);
     expect(wrapper).toMatchSnapshot();
-  })
-})
+  });
+});
