+++ src/components/common/Snackbar/SharedSnackbar.component.jsx
