+++ src/components/screens/Login/Login.test.js
@@ -1,6 +1,6 @@
 import React from "react";
 import {shallow} from 'enzyme';
-import Login from "../Login";
+import Login from "./Login";
 
 
 describe("<Login />", () => {
